# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ada21Tech AI-Powered Student Assessment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students/register:
    post:
      operationId: registerStudent
      tags: [students]
      summary: Register a new student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterStudentBody"
      responses:
        "201":
          description: Student registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict - email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /students/login:
    post:
      operationId: loginStudent
      tags: [students]
      summary: Student login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /students/me:
    get:
      operationId: getMyProfile
      tags: [students]
      summary: Get current student profile
      parameters:
        - in: query
          name: studentId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
...
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get a single student by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateStudent
      tags: [students]
      summary: Update student (admin can approve/change tier)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStudentBody"
      responses:
        "200":
          description: Updated student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /subjects:
    get:
      operationId: listSubjects
      tags: [subjects]
      summary: List all subjects
      responses:
        "200":
          description: List of subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubjectWithCount"
    post:
      operationId: createSubject
      tags: [subjects]
      summary: Create a new subject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubjectBody"
      responses:
        "201":
          description: Subject created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubjectWithCount"
  /subjects/{id}:
    delete:
      operationId: deleteSubject
      tags: [subjects]
      summary: Delete a subject
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Student:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        grade:
          type: string
        status:
          type: string
          enum: [pending, approved, suspended]
        subscriptionTier:
          type: string
          enum: [none, grade6, grade7, grade8, grade9, grade10, grade11, grade12]
        subscriptionPrice:
          type: ["integer", "null"]
        telebirrReceipt:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - grade
        - status
        - subscriptionTier
        - subscriptionPrice
        - telebirrReceipt
        - createdAt
    RegisterStudentBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        grade:
          type: string
        telebirrReceipt:
          type: ["string", "null"]
      required:
        - name
        - email
        - password
        - grade
    UpdateStudentBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, approved, suspended]
        subscriptionTier:
          type: string
          enum: [none, grade6, grade7, grade8, grade9, grade10, grade11, grade12]
        subscriptionPrice:
          type: ["integer", "null"]
        grade:
          type: ["string", "null"]
        telebirrReceipt:
          type: ["string", "null"]
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        role:
          type: string
          enum: [student]
        user:
          $ref: "#/components/schemas/Student"
      required:
        - token
        - role
        - user
...
    SubjectWithCount:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        grade:
          type: string
        description:
          type: ["string", "null"]
        questionCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - grade
        - description
        - questionCount
        - createdAt
    CreateSubjectBody:
      type: object
      properties:
        name:
          type: string
        grade:
          type: string
        description:
          type: ["string", "null"]
      required:
        - name
        - grade
...