# lib/api-zod/src/generated/types/types/registerStudentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ada21Tech AI-Powered Student Assessment Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterStudentBody {
  name: string;
  email: string;
  password: string;
  grade: string;
  /** @nullable */
  telebirrReceipt?: string | null;
}