# lib/api-zod/src/generated/types/types/student.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ada21Tech AI-Powered Student Assessment Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { StudentStatus } from "./studentStatus";
import type { StudentSubscriptionTier } from "./studentSubscriptionTier";

export interface Student {
  id: number;
  name: string;
  email: string;
  grade: string;
  status: StudentStatus;
  subscriptionTier: StudentSubscriptionTier;
  /** @nullable */
  subscriptionPrice: number | null;
  /** @nullable */
  telebirrReceipt: string | null;
  createdAt: Date;
}